feat: add constants module and clean duplicates
