+++ src/generator.py
@@ -34,9 +34,9 @@ from .loop_builder import (
 from .puzzle_io import save_puzzle
 from .validator import validate_puzzle, _has_zero_adjacent
 from .puzzle_builder import _build_puzzle_dict, _reduce_clues
+from .constants import MAX_SOLVER_STEPS
 
 from .types import Puzzle
-
 
 
 logging.basicConfig(
@@ -64,22 +64,7 @@ MIN_HINT_RATIO = {
 # リトライ回数を増やしてヒント削減に失敗しにくくする
 RETRY_LIMIT = 5
 
-# ソルバーが探索する最大ステップ数。超えると途中で打ち切る
-# ソルバーのステップ上限を増加させて解の探索精度を向上させる
-MAX_SOLVER_STEPS = 500000
-
-
-def _evaluate_difficulty(steps: int, depth: int) -> str:
-    """ソルバー統計から難易度を推定する関数"""
-
-    # 解析ステップ数とバックトラック深さを基準に難易度を決める
-    if steps < 1000 and depth <= 2:
-        return "easy"
-    if steps < 10000 and depth <= 10:
-        return "normal"
-    if steps < 100000 and depth <= 30:
-        return "hard"
-    return "expert"
+# MAX_SOLVER_STEPS と _evaluate_difficulty は constants モジュールへ移動した
 
 
 def generate_puzzle(
