+++ src/puzzle_builder.py
@@ -7,27 +7,14 @@ import random
 from typing import Any, Dict, List, Optional
 
 from .solver import PuzzleSize, count_solutions
+from .constants import MAX_SOLVER_STEPS, _evaluate_difficulty
 
 Puzzle = Dict[str, Any]
 
 # JSON スキーマのバージョン
 SCHEMA_VERSION = "2.0"
 
-# ソルバーのステップ上限
-MAX_SOLVER_STEPS = 500000
-
-
-def _evaluate_difficulty(steps: int, depth: int) -> str:
-    """ソルバー統計から難易度を推定する関数"""
-
-    # 解析ステップ数とバックトラック深さを基準に難易度を決定
-    if steps < 1000 and depth <= 2:
-        return "easy"
-    if steps < 10000 and depth <= 10:
-        return "normal"
-    if steps < 100000 and depth <= 30:
-        return "hard"
-    return "expert"
+# MAX_SOLVER_STEPS と _evaluate_difficulty は constants モジュールに移動
 
 
 def _reduce_clues(
